+++ src/optimization_metrics.cpp
@@ -17,15 +17,15 @@ namespace SparseLib::Metrics {
 	return c;
     }
 
-    double fbmDiagonalDistance(const Sparse::CSCMatrix& dsm) {
-    double total = 0.0;
+    long long fbmDiagonalDistance(const Sparse::CSCMatrix& dsm) {
+    long long total = 0;
     for (int col = 0; col < dsm.cols; col++) {
         const int start = dsm.col_ptrs[col];
         const int end   = dsm.col_ptrs[col + 1];
         for (int idx = start; idx < end; idx++) {
             const int row = dsm.row_indices[idx];
             if (row < col) {
-                total += static_cast<double>(col - row);
+                total += col - row;
             }
         }
     }
